requirements.txt
streamlit
google-generativeai
pandas
openpyxl
python-dotenv
sqlite3-binary